Ajustes finales  haciendo buen uso de encapsulamiento y definición de parámetros

+++ src/main/java/com/tyba/sqapractice/jise/sp331/Audiobook.java
@@ -54,7 +54,7 @@ public class Audiobook extends AudioTrack implements INavigation {
     }
 
     @Override
-    public void changeSpeed(int speed) {
+    public void changeSpeed(double speed) {
         System.out.println("Changing speed of audiobook: " + getTitle() + " to " + speed);
     }
 }
